- sample: Extrusion Indoor 3D - sample: Symbol Click Listener (missing long click)

+++ Naxam.Mapbox.Forms/Annotations/Annotation.Symbol.cs
@@ -38,19 +38,19 @@ namespace Naxam.Mapbox.Annotations
     {
         public LatLng Coordinates { get; set; }
 
-        public double? IconOpacity { get; set; }
+        public float? IconOpacity { get; set; }
 
-        public double? IconSize { get; set; }
+        public float? IconSize { get; set; }
 
-        public double? IconHaloBlur { get; set; }
+        public float? IconHaloBlur { get; set; }
 
-        public double? IconHaloWidth { get; set; }
+        public float? IconHaloWidth { get; set; }
 
         public IconImageSource IconImage { get; set; }
 
-        public double? IconRotate { get; set; }
+        public float? IconRotate { get; set; }
 
-        public double[] IconOffset { get; set; }
+        public float[] IconOffset { get; set; }
 
         public string IconAnchor { get; set; }
 
@@ -72,25 +72,25 @@ namespace Naxam.Mapbox.Annotations
 
         public Color? IconHaloColor { get; set; }
 
-        public double? TextSize { get; set; }
+        public float? TextSize { get; set; }
 
-        public double? TextMaxWidth { get; set; }
+        public float? TextMaxWidth { get; set; }
 
-        public double? TextLetterSpacing { get; set; }
+        public float? TextLetterSpacing { get; set; }
 
-        public double? TextRadialOffset { get; set; }
+        public float? TextRadialOffset { get; set; }
 
-        public double? TextRotate { get; set; }
+        public float? TextRotate { get; set; }
 
-        public double[] TextOffset { get; set; }
+        public float[] TextOffset { get; set; }
 
-        public double? TextOpacity { get; set; }
+        public float? TextOpacity { get; set; }
 
-        public double? TextHaloBlur { get; set; }
+        public float? TextHaloBlur { get; set; }
 
-        public double? TextHaloWidth { get; set; }
+        public float? TextHaloWidth { get; set; }
 
-        public double? SymbolSortKey { get; set; }
+        public float? SymbolSortKey { get; set; }
 
         public bool? IsDraggable { get; set; }
     }
